Move extensions and other test helpers into a separate file for ease of use in other test suites.
